==> 01_data_cleaning.ipynb <==
### Data Cleaning
#### The purpose of this notebook is to create cleaned .csv files to export for use in my data analyses

More information about this project is available in my github repo here: https://github.com/Noah-Baustin/sf_crime_data_analysis

#import modules
import pandas as pd

# import historical csv into a variable
historical_data = pd.read_csv('raw_data/SFPD_Incident_Reports_2003-May2018/Police_Department_Incident_Reports__Historical_2003_to_May_2018.csv', dtype=str)

# import newer csv into a variable
newer_data = pd.read_csv('raw_data/SFPD_Incident_Reports_2018-10.14.21/Police_Department_Incident_Reports__2018_to_Present(1).csv', dtype=str)

Trim the extra columns that we don't need from the historical data:

historical_data = historical_data[
    ['PdId', 'IncidntNum', 'Incident Code', 'Category', 'Descript',
       'DayOfWeek', 'Date', 'Time', 'PdDistrict', 'Resolution', 'X',
       'Y', 'location']
].reset_index(drop=True)

Change the column names in the historical data to match the API names in the newer data. The SFPD published a key that I used to translate the column names over, which can be found on pg two of this document: https://drive.google.com/file/d/13n7pncEOxFTWig9-sTKnB2sRiTB54Kb-/view?usp=sharing

historical_data.rename(columns={'PdId': 'row_id',
                                'IncidntNum': 'incident_number',
                                'Incident Code': 'incident_code',
                                'Category': 'incident_category',
                                'Descript': 'incident_description',
                                'DayOfWeek': 'day_of_week',
                                'Date': 'incident_date',
                                'Time': 'incident_time',
                                'PdDistrict': 'police_district',
                                'Resolution': 'resolution',
                                'X': 'longitude',
                                'Y': 'latitude',
                                'location': 'the_geom'
                               }, 
                       inplace=True)

historical_data

Now let's trim down the columns from the newer dataset so that we're only working with columns that match up to the old data. 

Note: there's no 'the geom' column, but the column 'point' is equivelant. 

newer_data = newer_data[
    ['Row ID', 'Incident Number', 'Incident Code', 'Incident Category', 
     'Incident Description', 'Incident Day of Week', 'Incident Date', 'Incident Time', 
     'Police District', 'Resolution', 'Longitude', 'Latitude', 'Point']
].copy()

Change the column names in the newer dataset to match the API names of the columns. Doing this because the original column names have spaces, which could cause issues down the road.

newer_data.rename(columns={'Row ID': 'row_id',
                           'Incident Number': 'incident_number',
                           'Incident Code': 'incident_code',
                           'Incident Category': 'incident_category',
                           'Incident Description': 'incident_description',
                           'Incident Day of Week': 'day_of_week', 
                           'Incident Date': 'incident_date',
                           'Incident Time': 'incident_time',
                           'Police District': 'police_district',
                           'Resolution': 'resolution',
                           'Longitude': 'longitude', 
                           'Latitude': 'latitude',
                           'Point': 'the_geom' 
                               }, 
                       inplace=True)

newer_data

historical_data.columns

newer_data.columns

Now that our datasets have matching columns, let's merge them together. 

frames = [historical_data, newer_data]
all_data = pd.concat(frames)

The dataframe all_data now contains our combined dataset!

all_data.info()
all_data.head()

We need to convert our incident_date column into a datetime format

all_data['incident_date'] = pd.to_datetime(all_data['incident_date'])

all_data['incident_date'].min()

all_data['incident_date'].max()

We can see from the date max and min that we've got our full set of date ranges from 2003 to 2021 in this new combined dataframe.

Since our string search we'll need to pull out our marijuana cases is cap sensitive, let's put all our values in the incident_description  and police district in lowercase:

all_data['incident_description'] = all_data['incident_description'].str.lower()

all_data['police_district'] = all_data['police_district'].str.lower()

Now let's create a dataframe with all our marijuana data:

all_data_marijuana = all_data[
    all_data['incident_description'].str.contains('marijuana')
].reset_index(drop=True)

all_data_marijuana

all_data_marijuana['incident_date'].min()

all_data_marijuana['incident_date'].max()

The incident dates show that we're getting marijuana incidents from 2003 all the way up to 2021. Great!

Now let's take a look at our incident description values:

all_data_marijuana['incident_description'].unique()

We can see there are slight differences in labeling of the same type of crime, likely cause by the transfer to the new system in 2018. So let's clean up the incidents column in both our dataframes.

all_data_marijuana = all_data_marijuana.replace({'incident_description' : 
                                            { 'marijuana, possession for sale' : 'possession of marijuana for sales',
                                              'marijuana, transporting' : 'transportation of marijuana',
                                              'marijuana, cultivating/planting' : 'planting/cultivating marijuana',
                                              'marijuana, sales' : 'sale of marijuana',
                                              'marijuana, furnishing' : 'furnishing marijuana',
                                            }
                                            })

all_data = all_data.replace({'incident_description' : 
                                            { 'marijuana, possession for sale' : 'possession of marijuana for sales',
                                              'marijuana, transporting' : 'transportation of marijuana',
                                              'marijuana, cultivating/planting' : 'planting/cultivating marijuana',
                                              'marijuana, sales' : 'sale of marijuana',
                                              'marijuana, furnishing' : 'furnishing marijuana',
                                            }
                                            })

all_data_marijuana['incident_description'].unique()

Looks good!

Now let's export our two dataframes to .csv's that we can now use in other data analysis!

all_data.to_csv("all_data.csv", index=False)
all_data_marijuana.to_csv("all_data_marijuana.csv", index=False)
