==> 02_marijuana_incidents_per_year.ipynb <==
### The purpose of this notebook is to answer question one of my analysis questions:

#### How many marijuana-involved incidents did the SFPD report each year over time?

#import modules
import pandas as pd
import altair as alt

Import our cleaned dataset that contains all of our marijuana incidents. We made this .csv file in the 01_data_cleaning notebook.

mari_incidents = pd.read_csv('all_data_marijuana.csv', dtype=str)

Convert our incident dates to a datetime data format.

mari_incidents['incident_date'] = pd.to_datetime(mari_incidents['incident_date'])

Check our date ranges

mari_incidents['incident_date'].min()

mari_incidents['incident_date'].max()

Looks like we've got a full year of data for 2003, our earliest year. But since 2021 ends in October, we can't do full annual analysis on that year. So let's make a dataframe with our full years of data.

full_years = mari_incidents[
    (mari_incidents['incident_date'] >= '2003-01-01') &
    (mari_incidents['incident_date'] < '2021-01-01')
].reset_index(drop=True)

We know from our data dictionary that there are multiple row entries for some individual incidents. But we also know that the incident_number will remain the same across all entries related to the same incident. So since we're just looking at how many incidents there were in each year, we can go ahead and drop all the duplicates in the incident_number column. (Data dictionary is here: https://datasf.gitbook.io/datasf-dataset-explainers/sfpd-incident-report-2018-to-present)

full_years_incidents = full_years.drop_duplicates(subset=['incident_number'])

Now we want to get annual counts of incidents using groupby

annual_mari_incidents = full_years_incidents.groupby([pd.Grouper(key='incident_date', axis=0, freq='A')]).count()

annual_mari_incidents

Now let's subset for just one column, incident_number, and reset the index so that incident data becomes a new column. 

annual_mari_incidents = annual_mari_incidents[['incident_number']].reset_index()

annual_mari_incidents

Now let's create a year column

annual_mari_incidents['year'] = annual_mari_incidents['incident_date'].dt.year

annual_mari_incidents

Now let's trim down our columns, rename them, and reorder them.

annual_mari_incidents.rename(columns={'incident_number': 'marijuana_incidents'}, inplace=True)
annual_mari_incidents

annual_mari_incidents_final = annual_mari_incidents[['year', 'marijuana_incidents']].copy()

annual_mari_incidents_final

### Time to visualize!

alt.Chart(annual_mari_incidents_final).mark_bar().encode(
    x='year:O',
    y='marijuana_incidents'
).properties(
    title='San Francisco Police: Annual Number of Marijuana Incidents'
)
