==> 03_types_of_incidents.ipynb <==
### The purpose of this notebook is to answer question two of my analysis questions:

#### Which types of marijuana crimes did SFPD report each year? Comparing the types of incidents over time.

#import modules
import pandas as pd
import altair as alt

Import our cleaned dataset that contains all of our marijuana incidents. We made this .csv file in the data_cleaning notebook.

mari_incidents = pd.read_csv('all_data_marijuana.csv', dtype=str)

Convert our incident dates to a datetime data format.

mari_incidents['incident_date'] = pd.to_datetime(mari_incidents['incident_date'])

Check our date ranges

mari_incidents['incident_date'].min()

mari_incidents['incident_date'].max()

Looks like we've got a full year of data for 2003, our earliest year. But since 2021 ends in October, we can't do full annual analysis on that year. So let's make a dataframe with our full years of data.

full_years = mari_incidents[
    (mari_incidents['incident_date'] >= '2003-01-01') &
    (mari_incidents['incident_date'] < '2021-01-01')
].reset_index(drop=True)

full_years['incident_description'].unique()

We can see here that the 'incident_description' column contains the information about which type of marijuana crime the police department logged in its incident database. And we can see from the unique entries in that column that each row only contains a single crime listed in the incident_description column. So while it's true that there might be multiple rows in our dataset that describe a single incident, we don't need to drop the duplicate incident numbers for this analysis. Because to capture all the marijuana crimes in a specific incident, there will be multiple rows for that incident. 

So now we're going to use groupby to count up the number of marijuana incidents in each category during the full duration of our data.

description_counts_all = full_years.groupby(['incident_description']).count()

Now let's isolate one column that we know won't have any null values: row_id

clean_description_counts_all = description_counts_all[['row_id']].copy()

clean_description_counts_all = clean_description_counts_all.reset_index()

#rename columns
clean_description_counts_all.columns = ['crime', 'number_of_incidents']

#sort by number of incidents
clean_description_counts_all = clean_description_counts_all.sort_values(by=['number_of_incidents'], ascending=False).reset_index(drop=True)

clean_description_counts_all

Great! We can draw some conclusions from this data. It shows us that from 2003 to 2020, the San Francisco Police Department responded to thousands of marijuana-related incidents. Possession of marijuana was the type of crime that the police department dealt with the most, followed by possession of marijuana for sales.

Here's a visualization of our data:

alt.Chart(clean_description_counts_all).mark_bar().encode(
    x=alt.X('crime:O', sort='-x'),
    y='number_of_incidents'
).properties(
    title='San Francisco Police: Number of Marijuana Incidents 2003-2020'
)

Now let's take a look at how the types of marijuana incidents actually changed year to year:

Create a dataframe that has the number of incidents for each type of crime for each year:

test = full_years.groupby(['incident_description', pd.Grouper(key='incident_date', axis=0, freq='A')]).count()

test

Clean up that dataframe:

test_2 = test['row_id'].reset_index()

test_2

Now I want to set up my dataframe so that the year is the row label and the type of crime is the column label and the values are the number of incidents.

incidents_per_year =  test_2.pivot(index='incident_date',
                       columns='incident_description',
                       values='row_id',
                      )

incidents_per_year

incidents_per_year = incidents_per_year.reset_index()

Let's clean up this table a bit:

incidents_per_year['year'] = incidents_per_year['incident_date'].dt.year

incidents_per_year_final = incidents_per_year[
    ['year', 
     'furnishing marijuana', 
     'marijuana offense', 
     'planting/cultivating marijuana',
     'possession of marijuana',
     'possession of marijuana for sales',
     'sale of marijuana',
     'transportation of marijuana'
    ]].copy()

incidents_per_year_final

#### There's our table! It shows us the number of incidents in the San Francisco Police Department database for each type of marijuana crime in each year from 2003 to 2020!

Generally speaking, the table looks good. One hiccup is that the 'marijuana offense' incident description didn't come into use until 2018, so we don't have a full history of data for that bucket. Also, an additional reporting question would be: did SFPD begin to code different types of marijuana crimes into the more general description 'marijuana offense' after 2018? For example, we're no longer seeing possession of marijuana crimes after 2019, but could those be showing up in this more general category?

#### Now let's visualize our data!

First, we've got to massage the shape of our data using the melt function.

chart_data = incidents_per_year_final.melt('year')

chart_data

Now that we've got our data in the correct shape, we can make a multi series line chart!

alt.Chart(chart_data).mark_line().encode(
    x='year:O',
    y='value',
    color='incident_description'
).properties(
    title='San Francisco Police: Annual Marijuana Incident Types'
)
