==> 04_incident_districts.ipynb <==
### The purpose of this notebook is to answer question three of my analysis questions:

#### How many marijuana-involved incidents did SFPD report different police districts across the city?

#import modules
import pandas as pd
import altair as alt

Import our cleaned dataset that contains all of our marijuana incidents. We made this .csv file in the data_cleaning notebook.

mari_incidents = pd.read_csv('all_data_marijuana.csv', dtype=str)

Convert our incident dates to a datetime data format.

mari_incidents['incident_date'] = pd.to_datetime(mari_incidents['incident_date'])

Check our date ranges

mari_incidents['incident_date'].min()

mari_incidents['incident_date'].max()

Looks like we've got a full year of data for 2003, our earliest year. But since 2021 ends in October, we can't do full annual analysis on that year. So let's make a dataframe with our full years of data.

full_years = mari_incidents[
    (mari_incidents['incident_date'] >= '2003-01-01') &
    (mari_incidents['incident_date'] < '2021-01-01')
].reset_index(drop=True)

We know from our data dictionary that there are multiple row entries for some individual incidents. But we also know that the incident_number will remain the same across all entries related to the same incident. So since we're just looking at how many incidents there were in each year in each district, we can go ahead and drop all the duplicates in the incident_number column:

full_years_incidents = full_years.drop_duplicates(subset=['incident_number'])

full_years_incidents.head()

incidents_by_district = full_years_incidents.groupby(['police_district']).count()

clean_incidents_by_district = incidents_by_district[['row_id']].copy()

clean_incidents_by_district = clean_incidents_by_district.reset_index()

#rename columns
clean_incidents_by_district.columns = ['police_district', 'number_of_incidents']

#sort by number of incidents
clean_incidents_by_district = clean_incidents_by_district.sort_values(by=['number_of_incidents'], ascending=False).reset_index(drop=True)

clean_incidents_by_district

So there we have it! That's all the marijuana related incidents the SF Police Department responded to from 2003-2020 by police district. It's clear that incidents that the police responded to are heavily weighted towards certain districts, including Southern, Tenderloin, and Park. An interesting follow up question would be to investigate why. Do more people live in those neighborhoods? Are more marijuana crimes committed in those neighborhoods? Do the police enforce marijuana laws differently in these neighborhoods than other parts of the city?

Let's visualize our data:

alt.Chart(clean_incidents_by_district).mark_bar().encode(
    x='police_district',
    y='number_of_incidents'
).properties(
    title='San Francisco Police: Marijuana Incidents by Police District 2003-2020'
)

That's the end of this analysis!
